update replace_first to just be a method of InitializedSExpression
